Modularizacion de LayerPane y ajustes varios

+++ paintPOO-dic-2024/src/backend/model/Figure.java
@@ -44,7 +44,7 @@ public abstract class Figure {
 
     public abstract void flipVertical();
 
-    public abstract Figure duplicate(int offset);
+    public abstract Figure duplicate(double offset);
 
     public abstract Figure[] divide();
 }
